+++ EDDiscovery/EliteDangerous/JournalEvents/JournalFetchRemoteModule.cs
@@ -18,14 +18,14 @@ namespace EDDiscovery.EliteDangerous.JournalEvents
         {
             StorageSlot = Tools.GetStringDef(evt["StorageSlot"]);
             StoredItem = Tools.GetStringDef(evt["StoredItem"]);
-            TransferCost = Tools.GetInt(evt["TransferCost"]);
+            TransferCost = Tools.GetLong(evt["TransferCost"]);
             Ship = Tools.GetStringDef(evt["Ship"]);
             ShipId = Tools.GetInt(evt["ShipID"]);
             ServerId = Tools.GetInt(evt["ServerId"]);
         }
         public string StorageSlot { get; set; }
         public string StoredItem { get; set; }
-        public int TransferCost { get; set; }
+        public long TransferCost { get; set; }
         public string Ship { get; set; }
         public int ShipId { get; set; }
         public int ServerId { get; set; }
