+++ EDDiscovery/EliteDangerous/JournalEvents/JournalModuleSell.cs
@@ -15,14 +15,14 @@ namespace EDDiscovery.EliteDangerous.JournalEvents
         {
             Slot = Tools.GetStringDef(evt["Slot"]);
             SellItem = Tools.GetStringDef(evt["SellItem"]);
-            SellPrice = Tools.GetInt(evt["SellPrice"]);
+            SellPrice = Tools.GetLong(evt["SellPrice"]);
             Ship = Tools.GetStringDef(evt["Ship"]);
             ShipId = Tools.GetInt(evt["ShipID"]);
 
         }
         public string Slot { get; set; }
         public string SellItem { get; set; }
-        public int SellPrice { get; set; }
+        public long SellPrice { get; set; }
         public string Ship { get; set; }
         public int ShipId { get; set; }
